Move Rect to bevy_ui and rename it to UiRect
