+++ crates/bevy_ui/src/flex/convert.rs
@@ -1,12 +1,12 @@
 use crate::{
     AlignContent, AlignItems, AlignSelf, Direction, Display, FlexDirection, FlexWrap,
-    JustifyContent, PositionType, Style, Val,
+    JustifyContent, PositionType, Style, UiRect, Val,
 };
-use bevy_math::{Rect, Size};
+use bevy_math::Size;
 
 pub fn from_rect(
     scale_factor: f64,
-    rect: Rect<Val>,
+    rect: UiRect<Val>,
 ) -> stretch::geometry::Rect<stretch::style::Dimension> {
     stretch::geometry::Rect {
         start: from_val(scale_factor, rect.left),
