+++ crates/bevy_ui/src/geometry.rs
@@ -1,0 +1,36 @@
+use bevy_reflect::Reflect;
+
+/// A rect, as defined by its "side" locations
+#[derive(Copy, Clone, PartialEq, Debug, Reflect)]
+#[reflect(PartialEq)]
+pub struct UiRect<T: Reflect + PartialEq> {
+    pub left: T,
+    pub right: T,
+    pub top: T,
+    pub bottom: T,
+}
+
+impl<T: Reflect + PartialEq> UiRect<T> {
+    pub fn all(value: T) -> Self
+    where
+        T: Clone,
+    {
+        UiRect {
+            left: value.clone(),
+            right: value.clone(),
+            top: value.clone(),
+            bottom: value,
+        }
+    }
+}
+
+impl<T: Default + Reflect + PartialEq> Default for UiRect<T> {
+    fn default() -> Self {
+        Self {
+            left: Default::default(),
+            right: Default::default(),
+            top: Default::default(),
+            bottom: Default::default(),
+        }
+    }
+}
