+++ crates/bevy_ui/src/ui_node.rs
@@ -1,6 +1,7 @@
+use crate::UiRect;
 use bevy_asset::Handle;
 use bevy_ecs::{prelude::Component, reflect::ReflectComponent};
-use bevy_math::{Rect, Size, Vec2};
+use bevy_math::{Size, Vec2};
 use bevy_reflect::{Reflect, ReflectDeserialize};
 use bevy_render::{
     color::Color,
@@ -89,13 +90,13 @@ pub struct Style {
     /// How items align according to the main axis
     pub justify_content: JustifyContent,
     /// The position of the node as descrided by its Rect
-    pub position: Rect<Val>,
+    pub position: UiRect<Val>,
     /// The margin of the node
-    pub margin: Rect<Val>,
+    pub margin: UiRect<Val>,
     /// The padding of the node
-    pub padding: Rect<Val>,
+    pub padding: UiRect<Val>,
     /// The border of the node
-    pub border: Rect<Val>,
+    pub border: UiRect<Val>,
     /// Defines how much a flexbox item should grow if there's space available
     pub flex_grow: f32,
     /// How to shrink if there's not enough space available
