+++ examples/game/game_menu.rs
@@ -85,7 +85,7 @@ mod splash {
             .spawn_bundle(ImageBundle {
                 style: Style {
                     // This will center the logo
-                    margin: Rect::all(Val::Auto),
+                    margin: UiRect::all(Val::Auto),
                     // This will set the logo to be 200px wide, and auto adjust its height
                     size: Size::new(Val::Px(200.0), Val::Auto),
                     ..default()
@@ -148,7 +148,7 @@ mod game {
             .spawn_bundle(NodeBundle {
                 style: Style {
                     // This will center the current node
-                    margin: Rect::all(Val::Auto),
+                    margin: UiRect::all(Val::Auto),
                     // This will display its children in a column, from top to bottom. Unlike
                     // in Flexbox, Bevy origin is on bottom left, so the vertical axis is reversed
                     flex_direction: FlexDirection::ColumnReverse,
@@ -166,7 +166,7 @@ mod game {
                 // Display two lines of text, the second one with the current settings
                 parent.spawn_bundle(TextBundle {
                     style: Style {
-                        margin: Rect::all(Val::Px(50.0)),
+                        margin: UiRect::all(Val::Px(50.0)),
                         ..default()
                     },
                     text: Text::with_section(
@@ -182,7 +182,7 @@ mod game {
                 });
                 parent.spawn_bundle(TextBundle {
                     style: Style {
-                        margin: Rect::all(Val::Px(50.0)),
+                        margin: UiRect::all(Val::Px(50.0)),
                         ..default()
                     },
                     text: Text {
@@ -394,7 +394,7 @@ mod menu {
         // Common style for all buttons on the screen
         let button_style = Style {
             size: Size::new(Val::Px(250.0), Val::Px(65.0)),
-            margin: Rect::all(Val::Px(20.0)),
+            margin: UiRect::all(Val::Px(20.0)),
             justify_content: JustifyContent::Center,
             align_items: AlignItems::Center,
             ..default()
@@ -404,7 +404,7 @@ mod menu {
             // This takes the icons out of the flexbox flow, to be positionned exactly
             position_type: PositionType::Absolute,
             // The icon will be close to the left border of the button
-            position: Rect {
+            position: UiRect {
                 left: Val::Px(10.0),
                 right: Val::Auto,
                 top: Val::Auto,
@@ -421,7 +421,7 @@ mod menu {
         commands
             .spawn_bundle(NodeBundle {
                 style: Style {
-                    margin: Rect::all(Val::Auto),
+                    margin: UiRect::all(Val::Auto),
                     flex_direction: FlexDirection::ColumnReverse,
                     align_items: AlignItems::Center,
                     ..default()
@@ -434,7 +434,7 @@ mod menu {
                 // Display the game name
                 parent.spawn_bundle(TextBundle {
                     style: Style {
-                        margin: Rect::all(Val::Px(50.0)),
+                        margin: UiRect::all(Val::Px(50.0)),
                         ..default()
                     },
                     text: Text::with_section(
@@ -524,7 +524,7 @@ mod menu {
     fn settings_menu_setup(mut commands: Commands, asset_server: Res<AssetServer>) {
         let button_style = Style {
             size: Size::new(Val::Px(200.0), Val::Px(65.0)),
-            margin: Rect::all(Val::Px(20.0)),
+            margin: UiRect::all(Val::Px(20.0)),
             justify_content: JustifyContent::Center,
             align_items: AlignItems::Center,
             ..default()
@@ -538,7 +538,7 @@ mod menu {
         commands
             .spawn_bundle(NodeBundle {
                 style: Style {
-                    margin: Rect::all(Val::Auto),
+                    margin: UiRect::all(Val::Auto),
                     flex_direction: FlexDirection::ColumnReverse,
                     align_items: AlignItems::Center,
                     ..default()
@@ -607,7 +607,7 @@ mod menu {
     ) {
         let button_style = Style {
             size: Size::new(Val::Px(200.0), Val::Px(65.0)),
-            margin: Rect::all(Val::Px(20.0)),
+            margin: UiRect::all(Val::Px(20.0)),
             justify_content: JustifyContent::Center,
             align_items: AlignItems::Center,
             ..default()
@@ -621,7 +621,7 @@ mod menu {
         commands
             .spawn_bundle(NodeBundle {
                 style: Style {
-                    margin: Rect::all(Val::Auto),
+                    margin: UiRect::all(Val::Auto),
                     flex_direction: FlexDirection::ColumnReverse,
                     align_items: AlignItems::Center,
                     ..default()
@@ -705,7 +705,7 @@ mod menu {
     ) {
         let button_style = Style {
             size: Size::new(Val::Px(200.0), Val::Px(65.0)),
-            margin: Rect::all(Val::Px(20.0)),
+            margin: UiRect::all(Val::Px(20.0)),
             justify_content: JustifyContent::Center,
             align_items: AlignItems::Center,
             ..default()
@@ -719,7 +719,7 @@ mod menu {
         commands
             .spawn_bundle(NodeBundle {
                 style: Style {
-                    margin: Rect::all(Val::Auto),
+                    margin: UiRect::all(Val::Auto),
                     flex_direction: FlexDirection::ColumnReverse,
                     align_items: AlignItems::Center,
                     ..default()
