+++ examples/ui/font_atlas_debug.rs
@@ -50,7 +50,7 @@ fn atlas_render_system(
                 image: texture_atlas.texture.clone().into(),
                 style: Style {
                     position_type: PositionType::Absolute,
-                    position: Rect {
+                    position: UiRect {
                         top: Val::Px(0.0),
                         left: Val::Px(512.0 * x_offset),
                         ..default()
