+++ examples/window/low_power.rs
@@ -169,7 +169,7 @@ pub(crate) mod test_setup {
                 style: Style {
                     align_self: AlignSelf::FlexStart,
                     position_type: PositionType::Absolute,
-                    position: Rect {
+                    position: UiRect {
                         top: Val::Px(5.0),
                         left: Val::Px(5.0),
                         ..Default::default()
